Fixing main function errors.

+++ LACPP-Assignment-1/main.cpp
@@ -8,13 +8,13 @@ using namespace std;
 float sum = 0.0;
 mutex m;
 
-float f(int x) {
-    return 4/(1+x*x);
+float f(float x) {
+    return 4.0 / (1.0 + x*x);
 }
 
 void area(float a,float b,float c) {
     m.lock();
-    sum += (a+b)*c/2;
+    sum += (a + b) * c / 2.0;
     m.unlock();
 }
 
@@ -26,9 +26,9 @@ int main(int argc, const char * argv[]) {
     
     vector<thread> th;
     
-    float h = 1/trapezes;
+    float h = 1.0 / (float)trapezes;
     
-    for (int i=0; i<1; i+=h) {
+    for (float i = 0.0; i < 1.0; i += h) {
         th.push_back(thread(area,f(i),f(i+h),h));
     }
     
@@ -39,4 +39,4 @@ int main(int argc, const char * argv[]) {
     cout << "Sum = " << sum << endl;
     
     return 0;
-}
+}
